+++ libs/charts/base/src/lib/abstract/prizm-charts-abstract.ts
@@ -41,7 +41,6 @@ export abstract class PrizmChartsAbstractComponent<
   get options(): Partial<OPTIONS> {
     return this.origin?.options as Partial<OPTIONS>;
   }
-
   @Input()
   @prizmDefaultProp()
   set theme(value: PrizmChartTheme) {