+++ libs/charts/base/src/lib/components/gauge/model.ts
@@ -1,0 +1,4 @@
+import { Gauge, GaugeOptions } from '@antv/g2plot';
+
+export type PrizmChartsGaugeOptions = Omit<GaugeOptions, 'percent'>;
+export type PrizmChartsGaugeOrigin = Gauge;